using only local storage to store selected experiments
