+++ src/src/pages/Metrics/components/MetricsBar/MetricsBar.tsx
@@ -36,7 +36,6 @@ function MetricsBar({
   onResetConfigData,
   onLiveUpdateConfigChange,
   onSelectExperimentNamesChange,
-  updateSelectedExperimentNames,
   onToggleAllExperiments,
 }: IMetricsBarProps): React.FunctionComponentElement<React.ReactNode> {
   const [popover, setPopover] = React.useState<string>('');
@@ -52,7 +51,6 @@ function MetricsBar({
     experimentsEngine.experimentsState((state) => state);
 
   React.useEffect(() => {
-    updateSelectedExperimentNames();
     experimentsEngine.fetchExperiments();
     return () => {
       experimentsEngine.destroy();
