+++ src/src/pages/Params/Params.tsx
@@ -117,7 +117,6 @@ const Params = ({
   onRowsVisibilityChange,
   onParamsScaleTypeChange,
   onSelectExperimentNamesChange,
-  updateSelectedExperimentNames,
   onToggleAllExperiments,
 }: IParamsProps): React.FunctionComponentElement<React.ReactNode> => {
   let scaleStates = getDefaultScaleStates(highPlotData);
@@ -199,7 +198,6 @@ const Params = ({
               liveUpdateConfig={liveUpdateConfig}
               onLiveUpdateConfigChange={onLiveUpdateConfigChange}
               onSelectExperimentNamesChange={onSelectExperimentNamesChange}
-              updateSelectedExperimentNames={updateSelectedExperimentNames}
               onToggleAllExperiments={onToggleAllExperiments}
               title={pageTitlesEnum.PARAMS_EXPLORER}
             />
