+++ src/src/pages/Scatters/Scatters.tsx
@@ -65,9 +65,6 @@ function Scatters(
               onSelectExperimentNamesChange={
                 props.onSelectExperimentNamesChange
               }
-              updateSelectedExperimentNames={
-                props.updateSelectedExperimentNames
-              }
               onToggleAllExperiments={props.onToggleAllExperiments}
               title='Scatters explorer'
             />
