+++ src/src/pages/Scatters/ScattersContainer.tsx
@@ -70,6 +70,10 @@ function ScattersContainer(): React.FunctionComponentElement<React.ReactNode> {
   }, [scattersData?.rawData]);
 
   React.useEffect(() => {
+    scattersAppModel.updateSelectedExperiments();
+  }, []);
+
+  React.useEffect(() => {
     scattersAppModel.initialize(route.params.appId);
     let appRequestRef: IApiRequest<void>;
     if (route.params.appId) {
@@ -179,9 +183,6 @@ function ScattersContainer(): React.FunctionComponentElement<React.ReactNode> {
         onResetConfigData={scattersAppModel.onResetConfigData}
         onSelectExperimentNamesChange={
           scattersAppModel.onSelectExperimentNamesChange
-        }
-        updateSelectedExperimentNames={
-          scattersAppModel.updateSelectedExperimentNames
         }
         onToggleAllExperiments={scattersAppModel.onToggleAllExperiments}
         onSelectOptionsChange={scattersAppModel.onSelectOptionsChange}
