+++ src/src/types/pages/metrics/Metrics.d.ts
@@ -131,7 +131,6 @@ export interface IMetricProps extends Partial<RouteChildrenProps> {
   onDensityTypeChange: (type: DensityOptions) => void;
   onMetricsSelectChange: (options: ISelectOption[]) => void;
   onSelectExperimentNamesChange: (experimentName: string) => void;
-  updateSelectedExperimentNames: () => void;
   onToggleAllExperiments: (experimentNames: string[]) => void;
   onSelectRunQueryChange: (query: string) => void;
   onRowsVisibilityChange: (metricKeys: string[]) => void;
