+++ src/src/types/pages/params/Params.d.ts
@@ -91,7 +91,6 @@ export interface IParamsProps extends Partial<RouteChildrenProps> {
   onColorIndicatorChange: () => void;
   onParamsSelectChange: (options: ISelectOption[]) => void;
   onSelectExperimentNamesChange: (experimentName: string) => void;
-  updateSelectedExperimentNames: () => void;
   onToggleAllExperiments: (experimentNames: string[]) => void;
   onSelectRunQueryChange: (query: string) => void;
   onGroupingSelectChange: (params: IOnGroupingSelectChangeParams) => void;
