+++ src/src/types/pages/scatters/Scatters.d.ts
@@ -94,7 +94,6 @@ export interface IScattersProps extends Partial<RouteChildrenProps> {
   onNotificationDelete: (id: number) => void;
   onResetConfigData: () => void;
   onSelectExperimentNamesChange: (experimentName: string) => void;
-  updateSelectedExperimentNames: () => void;
   onToggleAllExperiments: (experimentNames: string[]) => void;
   onSelectOptionsChange: (options: ISelectOption[]) => void;
   onSelectRunQueryChange: (query: string) => void;
